requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
autopep8==1.5.2
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
discord-webhook==0.13.0
discord.py @ git+https://github.com/Rapptz/discord.py@6381b5a8f45854f7b2aaafac1bd4ae7c1f67f494
Flask==1.1.2
googletrans==2.4.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.0
MarkupSafe==1.1.1
multidict==6.0.2
numpy==1.18.3
pycodestyle==2.5.0
requests==2.23.0
soupsieve==2.0
typing-extensions==4.2.0
urllib3==1.25.9
watchgod==0.6
Werkzeug==1.0.1
yarl==1.7.2
